requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   jsonlines
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via docling
boto3==1.38.2
    # via -r requirements.in
botocore==1.38.2
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via
    #   docling
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
dill==0.4.0
    # via multiprocess
docling==2.31.0
    # via -r requirements.in
docling-core==2.28.0
    # via
    #   -r requirements.in
    #   docling
    #   docling-ibm-models
    #   docling-parse
docling-ibm-models==3.4.2
    # via docling
docling-parse==4.0.1
    # via docling
easyocr==1.7.2
    # via docling
et-xmlfile==2.0.0
    # via openpyxl
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
filetype==1.2.0
    # via docling
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.30.2
    # via
    #   docling
    #   docling-ibm-models
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
imageio==2.37.0
    # via scikit-image
jinja2==3.1.6
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonlines==3.1.0
    # via docling-ibm-models
jsonref==1.1.0
    # via docling-core
jsonschema==4.23.0
    # via docling-core
jsonschema-specifications==2025.4.1
    # via jsonschema
latex2mathml==3.77.0
    # via docling-core
lazy-loader==0.4
    # via scikit-image
lxml==5.4.0
    # via
    #   docling
    #   python-docx
    #   python-pptx
markdown-it-py==3.0.0
    # via rich
marko==2.1.3
    # via docling
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpire==2.10.2
    # via semchunk
mpmath==1.3.0
    # via sympy
multiprocess==0.70.18
    # via mpire
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
ninja==1.11.1.4
    # via easyocr
numpy==2.2.5
    # via
    #   docling-ibm-models
    #   easyocr
    #   imageio
    #   opencv-python-headless
    #   pandas
    #   safetensors
    #   scikit-image
    #   scipy
    #   shapely
    #   tifffile
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
opencv-python-headless==4.11.0.86
    # via
    #   docling-ibm-models
    #   easyocr
openpyxl==3.1.5
    # via docling
packaging==25.0
    # via
    #   huggingface-hub
    #   lazy-loader
    #   scikit-image
    #   transformers
pandas==2.2.3
    # via
    #   docling
    #   docling-core
pillow==11.2.1
    # via
    #   docling
    #   docling-core
    #   docling-ibm-models
    #   docling-parse
    #   easyocr
    #   imageio
    #   python-pptx
    #   scikit-image
    #   torchvision
pluggy==1.5.0
    # via docling
pyclipper==1.3.0.post6
    # via easyocr
pydantic==2.11.3
    # via
    #   docling
    #   docling-core
    #   docling-ibm-models
    #   docling-parse
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via docling
pygments==2.19.1
    # via
    #   mpire
    #   rich
pylatexenc==2.10
    # via docling
pypdfium2==4.30.1
    # via docling
python-bidi==0.6.6
    # via easyocr
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-docx==1.1.2
    # via docling
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
python-pptx==1.0.2
    # via docling
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   docling-core
    #   easyocr
    #   huggingface-hub
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   docling
    #   huggingface-hub
    #   transformers
rich==14.0.0
    # via typer
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rtree==1.4.0
    # via docling
s3transfer==0.12.0
    # via boto3
safetensors==0.5.3
    # via
    #   docling-ibm-models
    #   transformers
scikit-image==0.25.2
    # via easyocr
scipy==1.15.2
    # via
    #   docling
    #   easyocr
    #   scikit-image
semchunk==2.2.2
    # via docling-core
setuptools==79.0.1
    # via
    #   torch
    #   triton
shapely==2.1.0
    # via easyocr
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
soupsieve==2.7
    # via beautifulsoup4
sympy==1.13.3
    # via torch
tabulate==0.9.0
    # via
    #   docling-core
    #   docling-parse
tifffile==2025.3.30
    # via scikit-image
tokenizers==0.21.1
    # via transformers
torch==2.7.0
    # via
    #   docling-ibm-models
    #   easyocr
    #   safetensors
    #   torchvision
torchvision==0.22.0
    # via
    #   docling-ibm-models
    #   easyocr
tqdm==4.67.1
    # via
    #   docling
    #   docling-ibm-models
    #   huggingface-hub
    #   mpire
    #   semchunk
    #   transformers
transformers==4.51.3
    # via
    #   docling-core
    #   docling-ibm-models
triton==3.3.0
    # via torch
typer==0.15.2
    # via
    #   docling
    #   docling-core
typing-extensions==4.13.2
    # via
    #   beautifulsoup4
    #   docling-core
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   referencing
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   botocore
    #   requests
xlsxwriter==3.2.3
    # via python-pptx
